refactor(connections): extract provider catalog
